requirements.txt
alembic==1.13.1
Flask_Migrate==4.0.7
Flask_SQLAlchemy==3.1.1
Flask_Cors==4.0.0
Flask_WTF==1.2.1
Flask_Login==0.6.3
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
waitress==3.0.0